+++ Tests/MultipasteCoreTests/main.swift
@@ -6,5 +6,6 @@ PreferencesTests.registerAll()
 SnippetMatcherTests.registerAll()
 SemanticVersionTests.registerAll()
 UpdateCheckerTests.registerAll()
+PasteboardAugmenterTests.registerAll()
 
 exit(Int32(runAllTests()))